CPU is working.
This took hours to complete, because of a strange thing happening with
my timings. It's like the registers were not delayed by a cycle. I tryed
a lot of different tricks to make it work. And I found a way that
would have worked for everything, except for reset, which would trigger
immediately on the negedge. That was acceptable...but I wasn't
understanding why it wasn't working.

I verified that my original very simple code in always blocks were
correct.

Turns out the issue was with the testbench itself. It used blocking
assignments to the instructions instead of non-blocking, which made
it look the the data was updating before the clock changed.
See:
https://stackoverflow.com/questions/11672239/why-is-my-d-flip-flop-not-waiting-for-the-positive-edge-of-the-clock